+++ usecase/find_account_by_id.go
@@ -26,7 +26,7 @@ type (
 	// Output data
 	FindAccountByIDOutput struct {
 		ID                   string                        `json:"id"`
-		AvailableCreditLimit float64                       `json:"available_credit_limit"`
+		AvailableCreditLimit int64                         `json:"available_credit_limit"`
 		Document             FindAccountByIDDocumentOutput `json:"document"`
 		CreatedAt            string                        `json:"created_at"`
 	}
